Patch to use identify a token by its slot description instead of slot number
Thanks to Huie-Ying Lee for the patch
http://www.opensc-project.org/pipermail/opensc-devel/2008-September/011317.html

======================================
Problem
======================================
Currently, the pam_pkcs11 module uses slot_num for the slot/token selection.
However, this is not safe because there is no guaranteed ordering of slots
returned from the PKCS#11 C_GetSlotList() function, according to the RSA
PKCS#11 specification.

On Solaris OS, the libpkcs11 library uses plug-ins to provide the slots.  An
administrator can install or uninstall a plug-in by using a utility command
called "cryptoadm" on Solaris.  Therefore, a particular slot may be associated
with different slot numbers on different systems or even on the same system on
different runs.

======================================
Solution
======================================
To address the above issue,  our solution is to replace the "slot_num" entry
with a "slot_description" entry in the "pam_pkcs11.conf" configuration file for
the slot/token selection as illustrated below.

- pam_pkcs11.conf -

pkcs11_module default {

 module = /usr/lib/libpkcs11.so;
 description = "Solaris PKCS#11 Cryptographic Framework library";
 #
 # Which slot to use. The slot is specified by the slot description.
 # For example, slot_description = "Sun Crypto Softtoken".
 #
 # An administrator can use the "cryotoadm list -v" command to find
 # all the available slots and their slot descriptions. For more
 # information, see the libpkcs11(3LIB) and cryptoadm(1m) man pages.
 #
 # The default value is "none" which means to use the first slot with
 # an available token.
 #
 slot_description = "none"

 ...

}

The slot_description field will be used to specify the slot to be used.  Its
value should be the slot label for the slot, which is basically the same as the
slotDescription string in the CK_SLOT_INFO structure returned from the
C_GetSlotInfo() function.

In the original slot_num design, when slot_num is 0, it means to use the first
slot with an available token.  To provide a similar functionality, an
administrator can specify slot_description to be "none", which also means to
use the first slot with an available token.