Improved error messages Thanks to Dominik Fischer for the patch
" It adds displaying error messages to the user via pam_prompt. Because in
GDM the messages disappear so quick, I've added a new option
"err_display_time". After a message is shown with pam_prompt, a
"sleep(err_display_time)" is called. This gives the user a chance to
read the message. Also I prepend every message with an error number,
because this number is easier to remember, if a user reports a problem.

The messages can be disabled via the "quiet" option. "

http://www.opensc-project.org/pipermail/opensc-devel/2011-March/016184.html