When searching LDAP, filter on the certificate
When searching LDAP for an entry for a user, incorporate the contents of
the certificate in the filter that we send to the directory server.  Now
that we let the directory server look for entries with the certificate
in them, we don't need to walk the list of certificates in the retrieved
entry to check for matches, because the server says they're there.