Add an LDAP "uid_attribute", use it to speed up
Add an option to name an attribute of a user entry in the directory
server which contains the user name.  If it's specified, search for
entries that just match the certificate, and read the user name from a
matching entry.  If not, fall back to the old method of iterating
through all user names to find an entry that matches both the user name
and the certificate.  The first method should be faster.