Add "attribute_map" to LDAP mapping
Add an "attribute_map" list setting.  Items in the list take the form
"ldap_attribute_name=cert_attribute", where "ldap_attribute_name" is an
attribute that we'll name in a search filter, and cert_attribute is one
of "cn", "subject", "kpn", "email", "upn", "uid", or "cert".

When searching the directory, we'll try the configured map attribute
pairs before falling back to the traditional map: the attribute named by
the "attribute" configuration setting should contain the certificate.