pkcs11_listcerts: do not fail on certificate error
If the verification of a certificates fails then display an error and
continue with the next certificate instead of exiting.

A certificate verification failure may be expected (certificate expired
or not able to find the CA root certificate). And the valid certificate
that we want to use may be the next in the list.