Fixed segfault and fetch problems when checking CRLs
Fixed segfault issue in src/common/cert_vfy.c that occurs when
an attempt is made to check a certificate's CRL.  This seems to
be caused by changes that happened in the OpenSSL API, and got
overlooked during updates to the code.

Also fixed a problem in src/common/uri.c in the builtin URI fetch
via HTTP where an extra newline (and missing carriage-returns) were
sent, causing the HTTP request to fail.